fix: register handler within init script
