+++ package.json
@@ -35,8 +35,9 @@
     "@nextcloud/moment": "^1.3.4",
     "@nextcloud/paths": "^2.2.1",
     "@nextcloud/router": "^3.0.1",
-    "@nextcloud/vue": "^8.27.0",
     "@nextcloud/sharing": "^0.2.4",
+    "@nextcloud/viewer": "^1.0.0",
+    "@nextcloud/vue": "^8.30.0",
     "vue": "^2.7.16",
     "vue-material-design-icons": "^5.3.1"
   },
