+++ webpack.js
@@ -10,6 +10,7 @@ const BabelLoaderExcludeNodeModulesExcept = require('babel-loader-exclude-node-m
 
 webpackConfig.entry = {
 	viewer: path.join(__dirname, 'src', 'viewer.js'),
+	'init-viewer': path.join(__dirname, 'src', 'init-viewer.js'),
 	fileActions: path.join(__dirname, 'src', 'file-actions.js'),
 	document: path.join(__dirname, 'src', 'document.js'),
 	admin: path.join(__dirname, 'src', 'admin.js'),
